Add option to not validate settings
This can be used to only read in valid settings (since we cannot read in invalid ones) but does not throw an error. Invalid values will then not overwrite the default ones, making it possible to still load the rest of valid config values in. This can then be used in other logics to recover settings and similar, to not fully brick autostart of the app in case auf faulty settings.